requirements/prod.txt
dj-database-url==0.4.2
Django==1.11.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
pytz==2017.2
raven==6.1.0
